+++ R/hotspot_dual_kde.R
@@ -42,6 +42,14 @@
 #' @param weights \code{NULL} (the default) or a vector of length two giving
 #'   either \code{NULL} or the name of a column in each of \code{x} and
 #'   \code{y} to be used as weights for weighted counts and KDE values.
+#' @param transform the underlying SpatialKDE package cannot calculate kernel
+#'   density for lon/lat data, so this must be transformed to use a projected
+#'   co-ordinate reference system. If this argument is \code{TRUE} (the 
+#'   default) and \code{sf::st_is_longlat(x)} is \code{TRUE}, \code{x}, \code{y} 
+#'   and \code{grid} (if provided) will be transformed automatically using 
+#'   \code{link{st_transform_auto}} before the kernel density is estimated and
+#'   transformed back afterwards. Set this argument to \code{FALSE} to suppress 
+#'   automatic transformation of the data.
 #' @param quiet if set to \code{TRUE}, messages reporting the values of any
 #'   parameters set automatically will be suppressed. The default is
 #'   \code{FALSE}.
@@ -128,6 +136,7 @@ hotspot_dual_kde <- function(
   method = "ratio",
   grid = NULL,
   weights = NULL,
+  transform = TRUE,
   quiet = FALSE,
   ...
 ) {
@@ -312,6 +321,7 @@ hotspot_dual_kde <- function(
       grid,
       bandwidth = bandwidth_x,
       bandwidth_adjust = bandwidth_adjust_x,
+      transform = transform,
       quiet = quiet,
       ...
     )
@@ -322,6 +332,7 @@ hotspot_dual_kde <- function(
       bandwidth = bandwidth_x,
       bandwidth_adjust = bandwidth_adjust_x,
       weights = weights_x,
+      transform = transform,
       quiet = quiet,
       ...
     )
@@ -334,7 +345,9 @@ hotspot_dual_kde <- function(
       grid,
       bandwidth = bandwidth_y,
       bandwidth_adjust = bandwidth_adjust_y,
-      quiet = quiet,
+      transform = transform,
+      # Prevent duplicate messages
+      quiet = TRUE,
       ...
     )
   } else {
@@ -344,7 +357,9 @@ hotspot_dual_kde <- function(
       bandwidth = bandwidth_y,
       bandwidth_adjust = bandwidth_adjust_y,
       weights = weights_y,
-      quiet = quiet,
+      transform = transform,
+      # Prevent duplicate messages
+      quiet = TRUE,
       ...
     )
   }